refactor(eng-analytics): split presentation serializers by domain
